+++ ts/sagas/startup.ts
@@ -44,7 +44,6 @@ import {
   startAndReturnIdentificationResult,
   watchIdentificationRequest
 } from "./identification";
-import { watchMessageLoadRequest } from "./messages/messages";
 import { updateInstallationSaga } from "./notifications";
 import { loadProfile, watchProfileUpsertRequestsSaga } from "./profile";
 import { authenticationSaga } from "./startup/authenticationSaga";
@@ -60,6 +59,7 @@ import { watchApplicationActivitySaga } from "./startup/watchApplicationActivity
 import { watchMessagesLoadOrCancelSaga } from "./startup/watchLoadMessagesSaga";
 import { loadMessageWithRelationsSaga } from "./startup/watchLoadMessageWithRelationsSaga";
 import { watchLogoutSaga } from "./startup/watchLogoutSaga";
+import { watchMessageLoadSaga } from "./startup/watchMessageLoadSaga";
 import { watchPinResetSaga } from "./startup/watchPinResetSaga";
 import { watchSessionExpiredSaga } from "./startup/watchSessionExpiredSaga";
 import { watchWalletSaga } from "./wallet";
@@ -228,7 +228,7 @@ function* initializeApplicationSaga(): IterableIterator<Effect> {
   yield fork(watchMessagesLoadOrCancelSaga, backendClient.getMessages);
 
   // Load messages when requested
-  yield fork(watchMessageLoadRequest, backendClient.getMessage);
+  yield fork(watchMessageLoadSaga, backendClient.getMessage);
 
   // Load message and related entities (ex. the sender service)
   yield takeEvery(
